setup/alias path and refactor import path
