+++ src/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Logger } from '@nestjs/common';
-import { AppService } from './app.service';
+import { AppService } from '@app/app.service';
 
 @Controller()
 export class AppController {
