+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { CoreModule } from './core/core.module';
-import { UsersModule } from './domains/users/users.module';
-import { AuthModule } from './domains/auth/auth.module';
+import { AppController } from '@app/app.controller';
+import { AppService } from '@app/app.service';
+import { CoreModule } from '@core/core.module';
+import { UsersModule } from '@domains/users/users.module';
+import { AuthModule } from '@domains/auth/auth.module';
 
 @Module({
   imports: [CoreModule, UsersModule, AuthModule],
