+++ src/domains/auth/auth.module.ts
@@ -1,13 +1,13 @@
 import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { LocalStrategy } from './strategies/local.strategy';
-import { LoginValidatorMiddleware } from './middleware/login-validator/login-validator.middleware';
+import { AuthController } from '@domains/auth/auth.controller';
+import { AuthService } from '@domains/auth/auth.service';
+import { LocalStrategy } from '@domains/auth/strategies/local.strategy';
+import { LoginValidatorMiddleware } from '@domains/auth/middleware/login-validator/login-validator.middleware';
 import { JwtModule } from '@nestjs/jwt';
-import jwt from 'src/config/jwt';
-import { JwtStrategy } from './strategies/jwt.strategy';
+import jwt from '@config/jwt';
+import { JwtStrategy } from '@domains/auth/strategies/jwt.strategy';
 import { APP_GUARD } from '@nestjs/core';
-import { JwtAuthGuard } from './guards/jwt.guard';
+import { JwtAuthGuard } from '@domains/auth/guards/jwt.guard';
 
 @Module({
   imports: [JwtModule.registerAsync(jwt.asProvider())],
