+++ src/domains/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
+import { UsersController } from '@domains/users/users.controller';
+import { UsersService } from '@domains/users/users.service';
 
 @Module({
   controllers: [UsersController],
