+++ tsconfig.json
@@ -10,6 +10,13 @@
     "sourceMap": true,
     "outDir": "./dist",
     "baseUrl": "./",
+    "paths": {
+      "@app/*": ["src/*"],
+      "@common/*": ["src/common/*"],
+      "@config/*": ["src/config/*"],
+      "@core/*": ["src/core/*"],
+      "@domains/*": ["src/domains/*"]
+    },
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
